refactor: CoreDataManager 이름 변경 -> CoreDataStack, DiaryCoreDataManager 이름 변경 -> CoreDataManager
