+++ Diary/CoreData/CoreDataStack.swift
@@ -1,0 +1,52 @@
+//
+//  CoreDataStack.swift
+//  Diary
+//
+//  Created by Rowan, Harry on 2023/05/05.
+//
+
+import Foundation
+import CoreData
+
+final class CoreDataStack {
+    static let shared = CoreDataStack()
+    
+    private init() { }
+    
+    private var persistentContainer: NSPersistentContainer = {
+        let container = NSPersistentContainer(name: "DiaryDAO")
+        
+        container.loadPersistentStores(completionHandler: { (_, error) in
+            if let error = error as NSError? {
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        
+        return container
+    }()
+    
+    var context: NSManagedObjectContext {
+        return persistentContainer.viewContext
+    }
+    
+    func fetch<T: NSManagedObject>(request: NSFetchRequest<T>) -> [T] {
+        do {
+            let fetchResult = try self.context.fetch(request)
+            return fetchResult
+        } catch {
+            print(error.localizedDescription)
+            return []
+        }
+    }
+    
+    func saveContext() {
+        if context.hasChanges {
+            do {
+                try context.save()
+            } catch {
+                let nserror = error as NSError
+                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+            }
+        }
+    }
+}
